### 2022-02-23 #4188 @editorialbot
[PRE REVIEW]: MitoHEAR: an R package for the estimation and downstream statistical analysis of the mitochondrial DNA heteroplasmy calculated from single-cell datasets
**Submitting author:** @gabrielelubatti (<a href="http://orcid.org/0000-0001-8495-8096">Gabriele Lubatti</a>)
**Repository:** https://github.com/ScialdoneLab/MitoHEAR
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.0
