### 2022-02-25 #4188 @arfon
@gabrielelubatti – given this is the third time this package has been submitted to JOSS, I'm going to put it through a scope review again to make sure the editorial team is happy with the state of the submission now.
