+++ README.md
@@ -4,7 +4,11 @@ MitoHEAR (**Mito**chondrial **HE**teroplasmy **A**nalyze**R**) is an R package t
 The package has been used in a recently published paper ([Lima *et al.*, 2021, Nature Metabolism](https://www.nature.com/articles/s42255-021-00422-7?proof=t)), where we revealed that cells with higher levels of heteroplasmy are eliminated by cell competition in mouse embryos and are characterized by specific gene expression patterns.
 
 ## Installation
-To install MitoHEAR, please run the following:
+You can install the released version of MitoHEAR from [CRAN](https://CRAN.R-project.org) with:
+
+```install.packages("MitoHEAR")```
+
+And the development version from [GitHub](https://github.com/) with:
 ```
 if (!requireNamespace("devtools", quietly = TRUE)) install.packages("devtools")
 devtools::install_github("https://github.com/ScialdoneLab/MitoHEAR/tree/master")
