### 2022-03-08 #4188 @csoneson
👋🏻 @BatoolMM, @waldeyr - would you be interested in reviewing this submission for JOSS?

> MitoHEAR: an R package for the estimation and downstream statistical analysis of the mitochondrial DNA heteroplasmy calculated from single-cell datasets
