### 2022-03-09 #4188 @csoneson
@BatoolMM - oh, no such hurry! 🙂  If you accept the review assignment, and once we have secured one more reviewer, I will open a separate review issue on GitHub, where you will have a checklist with the things that we ask our reviewers to look for. Once we're there the review can start and we would typically like to have an initial round of comments within two weeks or so (although JOSS reviews are often iterative, and it is often easier to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire submission).
