### 2022-03-09 #4188 @BatoolMM
Thank you for the clarification! It makes more sense now - my only experience with reviewing research software is `R Journal`, so thank you for elaborate explanation. This paper looks interesting - I'm happy to accept the review assignment!
