### 2022-03-14 #4188 @csoneson
👋🏻  @afrendeiro, @kelly-sovacool - would one of you be interested in reviewing this submission for JOSS?

> MitoHEAR: an R package for the estimation and downstream statistical analysis of the mitochondrial DNA heteroplasmy calculated from single-cell datasets
