### 2022-03-17 #4188 @csoneson
ping @waldeyr, @afrendeiro, @kelly-sovacool - would one of you be willing to review this submission for JOSS?

> MitoHEAR: an R package for the estimation and downstream statistical analysis of the mitochondrial DNA heteroplasmy calculated from single-cell datasets
