### 2022-03-22 #4188 @sahilseth
Hi @csoneson <https://github.com/csoneson>,

Sure happy to review it.

On Tue, Mar 22, 2022 at 2:22 AM Charlotte Soneson ***@***.***>
wrote:

> 👋🏻 @juanvillada <https://github.com/juanvillada>, @sahilseth
> <https://github.com/sahilseth> - would you be interested in reviewing
> this submission for JOSS?
>
> MitoHEAR: an R package for the estimation and downstream statistical
> analysis of the mitochondrial DNA heteroplasmy calculated from single-cell
> datasets
>
