### 2022-03-22 ScialdoneLab/MitoHEAR#1 @sahilseth
default branch
Could you create the master branch as default, as now it goes to `main` which is empty?

https://github.com/ScialdoneLab/MitoHEAR

Also, would be great if you could add the license file.

P.S.
This is in reference to the JOSS review here:
https://github.com/openjournals/joss-reviews/issues/4265
