### 2022-04-18 #4265 @BatoolMM
Did Anyone of the reviewers have a piece of luck in installing the package [MitoHEAR](https://github.com/ScialdoneLab/MitoHEAR)! I can't seem to install it in my local machine - I was going to open an issue in the repository but checking with @sahilseth and @juanvillada beforehand! It can be something with my local environment and R version!
