### 2022-04-19 #4265 @BatoolMM
I managed to install it from GitHub - checking now the tests and quality!
