### 2022-04-20 #4265 @juanvillada
Hi all, I managed to install MitoHEAR in Debian (linux kernel 5.10.104) with the latest R v4.1.3. I used a Docker container so I could track installation potential issues faced by users that have no pre-installed R libraries.

The issue I found was regarding the libraries required by MitoHEAR:

```
ERROR: dependencies ‘Biostrings’, ‘ComplexHeatmap’, ‘GenomicRanges’, ‘IRanges’, ‘rlist’, ‘Rsamtools’ are not available for package ‘MitoHEAR’

```

--------------------
Note: I had to make sure `curl` and `xml` were properly installed in the system to avoid further error messages. I am in Debian, so in my case I used:

```sh
apt-get install libcurl4-gnutls-dev
apt-get install libxml2-dev
```
--------------------

I then tried to install the dependencies:

```{r}
if (!require("BiocManager", quietly = TRUE))
    install.packages("BiocManager")

BiocManager::install("Biostrings")
BiocManager::install("ComplexHeatmap")
BiocManager::install("GenomicRanges")
BiocManager::install("rlist")
BiocManager::install("Rsamtools")
```

And I finally was able to successfully install MitoHEAR with:

```r
install.packages("MitoHEAR")
```
