### 2022-04-20 #4265 @juanvillada
# 2. On functionality

Most functionality could be reproduced. It is great the authors provided the reproducible notebooks, thanks!

Other comments:

## 2.1. Function: get_raw_counts_allele ⚠️

I could not reproduce the example for the function [get_raw_counts_allele](https://github.com/ScialdoneLab/MitoHEAR#get_raw_counts_allele). I think this example/test should be better automated. I can load the input and inspect the input data, I can load the expected output and inspect the output data; but  it is difficult to generate the output from the input with the instructions given in the documentation.

@gabrielelubatti there moght better options but I am thinking if it would be possible to host the required `.bam` and `.bai` files in a `.tar.gz` file? and then use the `R` functions `download.file()` and `untar()` to download and extract these data? I believe this can help potential users to have the testing data ready in a more automated way rather to having to download the files manually to test this function?

## 2.2. Function: get_heteroplasmy ✅

- Rewording
>starts from the output of get_raw_counts and performs a two step (...)

should be

>starts from the output of get_raw_counts*_allele* and performs a two step (...)

I was able to reproduce the example provided for this function.

# Vignettes

I took advantage of the availability of these cool vignettes/notebooks to further test the functionality of `MitoHEAR`, here are some notes:

## 2.3. Notebook:  cell_competition_mt_example_notebook.Rmd ✅⚠️

Please be explicit with the dependencies for this notebook, I found these issues:

⚠️ Error in plot_genome_coverage(biomart_file, fastaFile, chr_name, heteroplasmy_matrix) :
  Package karyoploteR and regioneR needed for this function to work. Please install them: BiocManager::install('karyoploteR') and BiocManager::install('regioneR')

⚠️ Error in dpt_test(heteroplasmy_matrix_ci[, relevant_bases], time, index_ci,  :
  Package gam needed for method==gam. Please install it: install.packages('gam')

⚠️ Error in plot_spider_chart(colnames(sum_matrix_qc)[1], cluster_ci, heteroplasmy_matrix_ci,  :
  Package fmsb needed for this function. Please install it: install.packages('fmsb')

## 2.4. Notebook: cell_competition_bulk_data_mt_example_notebook.Rmd ⚠️

⚠️ `bulk_data_competition[[5]]` is `NULL`

```{r}
index_bulk <- bulk_data_competition[[5]]
index_bulk
# NULL
```

## 2.5. Notebook: lineage_tracing_example_notebook.Rmd ⚠️

⚠️ Line 81: `sc_data[[5]]` is `NULL`

```{r}
index_sc <- sc_data[[5]]
index_sc
# NULL
```

⚠️ Line 130: `sc_data[[5]]` is `NULL`

```{r}
index_sc <- sc_data[[5]]
index_sc
# NULL
```

⚠️ Line 234: `sc_data[[5]]` is `NULL`

```{r}
index_sc <- sc_data[[5]]
index_sc
# NULL
```

## 2.6. Notebook: Ludwig_et_al_example_notebook.Rmd

⚠️ Line 73: `tfi_fig_2[[5]]` is `NULL`

```{r}
index_ci <- tfi_fig_2[[5]]
index_ci
# NULL
```

✅ Line 291: Here `donor_1_2[[5]]` is NOT `NULL`. I am wondering if this has something to do with the fact that in this case:
>"We don't execute the function *get_raw_counts_allele* here and  we directly load his output. "

```{r}
index_ci <- donor_1_2[[5]]
index_ci
# Juan's note: Output looks as expected
```

I was able to test this notebook til Line 356 (Unsupervised cluster analysis among cells based on allele frequency values). My session is killed after a while and I have not found the reason, but I guess is due to my RAM availabity, I will try this again with more RAM soon.
