### 2022-04-20 #4265 @juanvillada
# on references:

It would be nice to use consistent journal names. Some are abbreviated and some are not.

- Nat Cell Biol
- Nat Metab
- EMBO Rep
- Nat Rev Gen
