### 2022-04-20 #4265 @BatoolMM
Thank you @juanvillada - The License is also listed differently in two places: Artistic-2.0 in `DESCRIPTION ` whereas it says GPL-3 in the main GitHub repo.
