### 2022-04-20 #4265 @juanvillada
🐳

I have created a public Docker image with all the required dependencies in case we need it to further test `MitoHEAR`.

https://hub.docker.com/r/biovillada/r-base-mitohear
