### 2022-04-21 #4265 @gabrielelubatti
Dear @juanvillada  and @BatoolMM , thanks for your comments and feedback! I will work on these soon and let you know.
Regarding the problem with the fifth element of the output given by function get_heteroplasmy (that in some notebooks is NULL), this is something expected.
In the documentation for the get_heteroplasmy, in the section value, for the fifth element is written "Indices of the samples that cover a base, for all bases and samples that pass the two consequentially filtering steps; if all the samples cover all the bases, then index is NULL". So basically if the parameter "filtering" in function get_heteroplasmy is equal to 1, then the fifth element of the output is NULL, if filtering is equal to 2, then the fifth element is not NULL.
