+++ README.md
@@ -14,6 +14,19 @@ MitoHEAR (**Mito**chondrial **HE**teroplasmy **A**nalyze**R**) is an R package t
 The package has been used in a recently published paper ([Lima *et al.*, 2021, Nature Metabolism](https://www.nature.com/articles/s42255-021-00422-7?proof=t)), where we revealed that cells with higher levels of heteroplasmy are eliminated by cell competition in mouse embryos and are characterized by specific gene expression patterns.
 
 ## Installation
+
+Before installing MitoHEAR, the following libraries should be installed
+```
+if (!require("BiocManager", quietly = TRUE))
+    install.packages("BiocManager")
+
+BiocManager::install("Biostrings")
+BiocManager::install("ComplexHeatmap")
+BiocManager::install("GenomicRanges")
+BiocManager::install("rlist")
+BiocManager::install("Rsamtools")
+```
+
 You can install the released version of MitoHEAR from [CRAN](https://CRAN.R-project.org) with:
 
 ```install.packages("MitoHEAR")```
