+++ Draft/paper.bib
@@ -59,13 +59,13 @@ keywords = {lineage tracing, single cell genomics, somatic mutations, mitochondr
 	doi = {10.1038/s42255-021-00422-7},
 	URL = {https://www.nature.com/articles/s42255-021-00422-7},
 	eprint = {https://www.nature.com/articles/s42255-021-00422-7.pdf},
-	journal = {Nature Metabolism}
+	journal = {Nat Metab}
 }
 
 @article{Nissanka2020,
 author = {Nissanka, Nadee and Moraes, Carlos T},
 title = {Mitochondrial DNA heteroplasmy in disease and targeted nuclease-based therapeutic approaches},
-journal = {EMBO reports},
+journal = {EMBO Rep},
 volume = {21},
 number = {3},
 pages = {e49612},
