+++ DESCRIPTION
@@ -5,7 +5,7 @@ Version: 0.1.0
 Author: Gabriele Lubatti
 Maintainer: Gabriele Lubatti <gabriele.lubatti@helmholtz-muenchen.de>
 Description: R package that allows the estimation and downstream statistical analysis of the mitochondrial DNA Heteroplasmy calculated from single-cell datasets.
-License: Artistic-2.0
+License: GPL-3
 Depends:
     R (>= 4.0)
 Imports:
