+++ R/get_heteroplasmy.R
@@ -50,7 +50,7 @@
 #'
 #' \item{index}{Indices of the samples that
 #' cover a base, for all bases and samples that pass the two consequentially
-#' filtering steps; if all the samples cover all the bases, then \emph{index} is NULL }
+#' filtering steps (if \emph{filtering} = 2); if all the samples cover all the bases (that is the case for \emph{filtering} = 1), then \emph{index} is NULL }
 #' @author Gabriele Lubatti \email{gabriele.lubatti@@helmholtz-muenchen.de}
 #'
 #' @export get_heteroplasmy
