### 2022-04-21 #4265 @gabrielelubatti
Regarding the comments from @juanvillada:
**On functionality**
1) I modified the example for the function [get_raw_counts_allele](https://github.com/ScialdoneLab/MitoHEAR#get_raw_counts_allele), using download.file() and untar() functions in order to make the test  more automated.
**Vignettes**
1) For the notebook "cell_competition_mt_example_notebook.Rmd" I added a sentence and a chunk of code at the beginning of the script, specifying the dependencies.
2) For the problem with the fifth element of the output given by function get_heteroplasmy, in addition to what I wrote in my previous comment, I slightly changed the documentation of get_heteroplasmy, in order to make things clearer.
**On references**
1) Regarding the references, I abbreviated them in order to keep consistency.
