### 2022-04-21 #4265 @gabrielelubatti
Regarding the comments from @BatoolMM:
I changed the license in the DESCRIPTION file, in order to be consistent with the one in the main GitHub repo (GPL-3).
