+++ DESCRIPTION
@@ -40,3 +40,4 @@ Encoding: UTF-8
 LazyData: true
 Config/testthat/edition: 3
 RoxygenNote: 7.1.1
+BugReports: https://github.com/ScialdoneLab/MitoHEAR/issues
