### 2022-04-21 #4265 @gabrielelubatti
Dear @juanvillada, thanks a lot  for your feedback and comments!
Regarding the last two points you mentioned:

-  **Please add the Community guidelines**
I added a section "Contributions and Support" in the README file of the GitHub repo. I also added a BugReports link in the DESCRIPTION file of the library

- ** Probably my last comment, and this is only a suggestion:**
Thanks for your suggestion, I will work on it soon
