### 2022-04-23 #4265 @BatoolMM
Thank you @gabrielelubatti for the changes you made to the installation guidelines and functionality. It's way better now. Thank you @juanvillada for all your suggestions.

I love the package but have a few comments about the text of the paper, vignettes to make it more user-friendly.

In terms of the paper:
- Can you elaborate more about how this software compares to other commonly-used packages or pipelines, maybe such as MToolBox (it doesn't have to be R packages)?
- Can you also fix the reference? Some of it seems off!
> Our package MitoHEAR covers all steps of the analysis, instead of focussing only on a few specific steps (as in, e.g., (Huang and Huang 2021),(Prashant et al. 2021)): from the processing of raw FASTQ files to the estimation of the heteroplasmy and downstream statistical analysis, with user-friendly functions that are highly customizable.
- I like how you describe the main functions, but can we have the last paragraph (see below) as bulletin points and maybe also include an example for the visualisation.
> Among the downstream analyses implemented in the package there are: 1. Several statistical tests (e.g., Wilcoxon rank-sum test) for the identification of the mtDNA positions with the most different levels of heteroplasmy between discrete groups of cells or along a trajectory of cells (i.e., cells sorted according to a diffusion pseudo time); 2. Plotting functions for the visualization of heteroplasmy and the corresponding allele frequency values among cells; 3. Hierarchical clustering of cells based on a distance matrix defined from the angular distance of allele frequencies that could be relevant for lineage tracing analysis (Ludwig et al. 2019).

As @juanvillada mentioned,  we appreciate the effort the authors put into the reproducible notebooks but in order to run the notebooks, I need to install them and knit them, make sure I have installed all dependencies - which might be straightforward - but for biologists who may not be familiar with R Markdown or want to quickly skim the notebooks with minimal effort, I recommend using `pkgdown`. Is there a reason that you prefer not to use it? It's not necessary but will make the package more user-friendly. You'll only need to run two functions `usethis::use_pkgdown()` and `pkgdown::build_site()` and you will have a site, which will have all the notebooks and can be added to the README.

Another few comments about the README, is it possible to avoid any hardcoded path? In the 1st line, you mention
```r
current_wd <- "/Users/gabriele.lubatti/Documents/test_bam/"
```
Although you already instruct the user to create their own working dir, it might be better to use - something like:

```r
current_wd <- getwd()

```
Also, the example input of `get_raw_counts_allele ` is written in 19 lines (one chunk), which I don't recommend for easy readability.  Would it be possible to divide it into chunks with more annotation and try to use the [highlights functionality](https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/creating-and-highlighting-code-blocks) for the code?

These are all simple suggestions to make the package more user friendly but are not necessarily needed to be added to the package. I didn't want this comment to be long - I'll move then to add a few other comments about the testing. I use usually `autotest` to test the function, which you can also try, if you want!
