+++ README.md
@@ -193,6 +193,7 @@ The heteroplasmy and the corresponding allele frequencies for the most relevant 
 plot_heteroplasmy(names(p_value_wilcox_test_sort)[1], heteroplasmy_matrix_ci, cluster_ci, index_ci)
 plot_allele_frequency(i, heteroplasmy_matrix_ci, allele_matrix_ci, cluster_ci, name_position_qc, name_position_allele_qc, 5, index_ci)
 ```
+<img src="https://github.com/ScialdoneLab/MitoHEAR/blob/master/Draft/docs/img/paper_fig_1.png" width="800" height="500">
 
 If for each sample a diffusion pseudo time information is available, then it is possible to detect the bases whose heteroplasmy changes in a significant way along pseudo-time with **dpt_test** and to plot the trend with **plot_dpt**.
 ```
@@ -201,7 +202,9 @@ dpt_analysis <- dpt_test(heteroplasmy_matrix_ci[, relevant_bases], time, index_c
 
 plot_dpt(dpt_analysis$Position[1], heteroplasmy_matrix_ci, cluster_ci, time, dpt_analysis, index_ci)
 ```
-It is also possible to perform a cluster analysis on the samples based on distance matrix obtained from allele frequencies with **clustering_angular_distance** and to visualize an heatmap of the distance matrix with samples sorted according to the cluster result with **plot_heatmap**. This approach could be useful for lineage tracing analysis. 
+<img src="https://github.com/ScialdoneLab/MitoHEAR/blob/master/Draft/docs/img/paper_fig_2.png" width="800" height="500">
+
+It is also possible to perform a cluster analysis on the samples based on distance matrix obtained from allele frequencies with **clustering_angular_distance** and to visualize an heatmap of the distance matrix with samples sorted according to the cluster result with **plot_distance_matrix**. This approach could be useful for lineage tracing analysis. 
 The data shown in the example below is bulk RNA seq mouse data from two mtDNA cell lines labelled *Loser* and *Winner* ([Lima *et al.*, Nature Metabolism, 2021 ](https://www.nature.com/articles/s42255-021-00422-7?proof=t))
 
 Pre-processing of the dataset following the same steps described in the section **Getting started**
@@ -229,7 +232,7 @@ cluster_bulk <- as.character(meta_data_ana_final_small[row.names(heteroplasmy_ma
 index_bulk <- bulk_data_competition[[5]]
 ```
 
-Unsupervised hierarchical clustering on the samples based on a distance matrix with the function **clustering_dist_ang**
+Unsupervised hierarchical clustering on the samples based on a distance matrix with the function **clustering_dist_ang**.
 
 ```
 result_clustering_sc <- clustering_angular_distance(heteroplasmy_matrix_bulk, allele_matrix_bulk, cluster_bulk, length(colnames(heteroplasmy_matrix_bulk)), deepSplit_param = 1, minClusterSize_param = 8, 0.2, min_value = 0.001, index = index_bulk, relevant_bases = NULL)
