+++ README.md
@@ -93,7 +93,7 @@ The output of **get_raw_counts_allele** is a list with three elements:
 1. **matrix_allele_counts**: matrix with rows equal to **cell_names** and with columns equal to the bases in the **path_fasta** file with the four possible alleles. For each pair sample-base there is the information about the counts on the alleles A,C,G and T
 2. **name_position_allele**: character vectors with length equal to the number of columns in **matrix_allele_counts** with information about the name of the bases and the corresponding allele
 3. **name_position**: character vectors with information about the name of the bases.
-````
+
 ```
 # development version from GitHub
 load(system.file("extdata", "output_SNP_mt.Rda", package = "MitoHEAR"))
@@ -117,7 +117,7 @@ name_position <- output_SNP_mt[[3]]
 name_position[1:8]
 [1] "1_MT" "1_MT" "1_MT" "1_MT" "2_MT" "2_MT" "2_MT" "2_MT"
 ```
-````
+
 
 
 ### get_heteroplasmy
