+++ README.md
@@ -281,7 +281,7 @@ For more exhaustive information about the functions offered by **MitoHEAR** see 
 ## Vignettes
 
 The following vignettes are provided within the package **MitoHEAR** (development version from [GitHub](https://github.com/)) and are accessible within R:
-
+For a complete overwiew of the vignettes without the need of using R, please [click here](https://scialdonelab.github.io/MitoHEAR/)
 
 ### **[cell_competition_mt_example_notebook.Rmd](https://github.com/ScialdoneLab/MitoHEAR/blob/master/vignettes/cell_competition_mt_example_notebook.Rmd):**
 ```
