+++ Draft/paper.bib
@@ -91,3 +91,17 @@ year = {2020}
 	doi = {10.1038/s41556-017-0017-8},
 	journal = {Nat Cell Biol}
 }
+@article{Calabrese2014,
+    author = {Calabrese, Claudia and Simone, Domenico and Diroma, Maria Angela and Santorsola, Mariangela and Guttà, Cristiano and Gasparre, Giuseppe and Picardi, Ernesto and Pesole, Graziano and Attimonelli, Marcella},
+    title = "{MToolBox: a highly automated pipeline for heteroplasmy annotation and prioritization analysis of human mitochondrial variants in high-throughput sequencing}",
+    journal = {Bioinformatics},
+    volume = {30},
+    number = {21},
+    pages = {3115-3117},
+    year = {2014},
+    month = {07},
+    issn = {1367-4803},
+    doi = {10.1093/bioinformatics/btu483},
+    url = {https://doi.org/10.1093/bioinformatics/btu483},
+    eprint = {https://academic.oup.com/bioinformatics/article-pdf/30/21/3115/17146381/btu483.pdf},
+}
