+++ Draft/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'MitoHEAR'
+title: 'MitoHEAR: an R package for the estimation and downstream statistical analysis of the mitochondrial DNA heteroplasmy calculated from single-cell datasets'
 date: "17 June 2021"
 bibliography: paper.bib
 authors:
@@ -27,26 +27,35 @@ tags:
 To produce the energy they need, eukaryotic cells rely on mitochondria, organelles that are equipped with their own DNA (mtDNA). Each cell includes multiple mtDNA copies that are not perfectly identical but have differences in their sequence; such sequence variability is called heteroplasmy.
 mtDNA heteroplasmy has been associated with diseases [@Nissanka2020], can affect cellular fitness and have an impact on cellular competition [@Lima2020].
 Several single-cell sequencing protocols provide the data to estimate mtDNA heteroplasmy, including single-cell DNA-seq, RNA-seq and ATAC-seq, in addition to dedicated protocols like MAESTER [@Miller2021].
-Here, we provide MitoHEAR (Mitochondrial HEteroplasmy AnalyzeR), a user-friendly software written in R that allows the estimation as well as downstream statistical analysis of the mtDNA heteroplasmy calculated from single-cell datasets. MitoHEAR starts from FASTQ files, computes the frequency of each allele and, starting from these, estimates the mtDNA heteroplasmy at each covered position for each cell.  
-The parameters of the analysis (e.g., the filtering of the mtDNA positions based on read quality and coverage) are easily tuneable in a very user-friendly way. Moreover, statistical tests are available to explore the dependency of the mtDNA heteroplasmy on continuous or discrete cell covariates (e.g., culture conditions, differentiation states, etc), as extensively shown in the detailed tutorials we include. 
+Here, we provide MitoHEAR (Mitochondrial HEteroplasmy AnalyzeR), a user-friendly software written in R that allows the estimation as well as downstream statistical analysis of the mtDNA heteroplasmy calculated from single-cell datasets. MitoHEAR takes as input BAM files, computes the frequency of each allele and, starting from these, estimates the mtDNA heteroplasmy at each covered position for each cell.  
+The parameters of the analysis (e.g., the filtering of the mtDNA positions based on read quality and coverage) are easily tuneable. Moreover, statistical tests are available to explore the dependency of the mtDNA heteroplasmy on continuous or discrete cell covariates (e.g., culture conditions, differentiation states, etc), as extensively shown in the detailed tutorials we include. 
 
 
 # Statement of need
-Despite mtDNA heteroplasmy has important consequences on human health [@Stewart2015] and embryonic development [@Floros2019], there are still many open questions on how heteroplasmy affects cell's ability to function and regarding the mechanisms cells use to keep it under control. 
+Despite mtDNA heteroplasmy has important consequences on human health [@Stewart2015] and embryonic development [@Floros2019], there are still many open questions on how heteroplasmy affects cells' ability to function and how cells keep it under control. 
 With the increasing availability of single-cell data, many questions can begin to be answered, but it is fundamental to have efficient and streamlined computational tools enabling researchers to estimate and analyse mtDNA heteroplasmy. 
-Our package MitoHEAR covers all steps of the analysis, instead of focussing only on a few specific steps (as in, e.g., [@Huang2021],[@Prashant2020]): from the processing of raw FASTQ files to the estimation of the heteroplasmy and downstream statistical analysis, with user-friendly functions that are highly customizable.  
+Existing packages (e.g., [@Huang2021],[@Prashant2020],[@Calabrese2014]) focus only on the first step consisting in the quantification of heteroplasmy from BAM files, but they do not provide any specific tools for further statistical analyses or plotting.
[…]
