### 2022-04-28 #4265 @gabrielelubatti
Dear @BatoolMM , thanks for your feedbacks and comments!

Here are some point-by-point answers to your questions:

In terms of the paper:
1) Can you elaborate more about how this software compares to other commonly-used packages or pipelines, maybe such as MToolBox (it doesn't have to be R packages)?

The other packages that are currently available focus on the first step of the analysis, consisting in the quantification of the heteroplasmy from the BAM files. However, they do not provide any additional tools or functions to perform downstream statistical analyses or plotting.
I have added a sentence on this point (see Lines 33 - 42).

2) Can you also fix the reference? Some of it seems off!

I double-checked and the references (especially those in the text you quote) and they seem fine. I don't know whether you noticed something in particular?

3) I like how you describe the main functions, but can we have the last paragraph (see below) as bulletin points and maybe also include an example for the visualisation.

Thanks for the suggestions! I've now structured the last paragraph as bullet points and I added three figures as example of the main downstream analysis offered by MitoHEAR.

4) Regarding the use of pkgdown

Thanks for the useful suggestion. I didn't know about this nice library. I have run the two functions you mentioned and I reported the link of the website on the README file of the GitHub repo (at the very beginning and at the starting point of the section Vignettes). Now it is definitely more user-friendly having a look at the tutorials!

5) Regarding the example input of get_raw_counts_allele

I have divided it into chunks using highlights functionality for highlighting the code from comments, as suggested, and I have removed hardcoded path using getwd() instead.

6) Regarding the testing part of the functions

Actually, I have used the library testthat (that should be quite similar to the autotest library) for testing some of the main functions of MitoHEAR.
