### 2022-05-08 ScialdoneLab/MitoHEAR#2 @BatoolMM
Increase the readability of the code in the README
Thank you for all the changes you made to the repo (openjournals/joss-reviews#4265), I love the changes you made to the README .

Just a simple tip when adding code to the READEM - in order to increase the readability, we recommend highlighting the syntax by defining the language in code block ( e.g. ```r )

For example

```
current_wd <- getwd()
url <- "https://hmgubox2.helmholtz-muenchen.de/index.php/s/7P9C57RxfKnH5Qx/download/input_bam_files.tar.gz"
destfile <- paste0(current_wd, "input_bam_files.tar.gz")
download.file(url, destfile, quiet = FALSE)
untar(destfile, exdir=current_wd)

load(system.file("extdata", "after_qc.Rda", package = "MitoHEAR"))
cell_names <- as.vector(after_qc$new_name)
cell_names <- cell_names[1:5]
cell_names[1:5]
[1] "24538_8_14" "24538_8_23" "24538_8_39" "24538_8_40" "24538_8_47"
```

will be:

```r
current_wd <- getwd()
url <- "https://hmgubox2.helmholtz-muenchen.de/index.php/s/7P9C57RxfKnH5Qx/download/input_bam_files.tar.gz"
destfile <- paste0(current_wd, "input_bam_files.tar.gz")
download.file(url, destfile, quiet = FALSE)
untar(destfile, exdir=current_wd)

load(system.file("extdata", "after_qc.Rda", package = "MitoHEAR"))
cell_names <- as.vector(after_qc$new_name)
cell_names <- cell_names[1:5]
cell_names[1:5]
[1] "24538_8_14" "24538_8_23" "24538_8_39" "24538_8_40" "24538_8_47"
```
 by only adding `r` in the code block!

It is not necessarily added, but it makes the code more readable!

An example of some editing to the README file with a table of content can be seen in [my fork here](https://github.com/BatoolMM/MitoHEAR)

Thank you again for your work!
