+++ README.md
@@ -7,6 +7,16 @@
 
 <!-- badges: end -->
 
+- [MitoHEAR](#mitohear)
+  * [Installation](#installation)
+  * [Getting started](#getting-started)
+    + [get_raw_counts_allele](#get-raw-counts-allele)
+    + [get_heteroplasmy](#get-heteroplasmy)
+  * [Down-stream analysis](#down-stream-analysis)
+  * [Vignettes](#vignettes)
+  * [Contributions and Support](#contributions-and-support)
+
+<small><i><a href='http://ecotrust-canada.github.io/markdown-toc/'>Table of contents generated with markdown-toc</a></i></small>
 
 # MitoHEAR
 [MitoHEAR](https://scialdonelab.github.io/MitoHEAR) (**Mito**chondrial **HE**teroplasmy **A**nalyze**R**) is an R package that allows the estimation as well as downstream statistical analysis of the mtDNA heteroplasmy calculated from single-cell datasets.
@@ -17,20 +27,20 @@ The package has been used in a recently published paper ([Lima *et al.*, 2021, N
 
 You can install the released version of MitoHEAR from [CRAN](https://CRAN.R-project.org) with:
 
-```
+```r
 if (!require("BiocManager", quietly = TRUE))
     install.packages("BiocManager")
 BiocManager::install("MitoHEAR")
 ```
 
 And the development version from [GitHub](https://github.com/) with:
-```
+```r
 if (!requireNamespace("devtools", quietly = TRUE)) install.packages("devtools")
 devtools::install_github("https://github.com/ScialdoneLab/MitoHEAR/tree/master")
 library(MitoHEAR)
 ```
 For installing also the vignettes provided within the package, please run the following:
-```
+```r
 if (!requireNamespace("devtools", quietly = TRUE)) install.packages("devtools")
 devtools::install_github("https://github.com/ScialdoneLab/MitoHEAR/tree/master", build_vignettes = TRUE)
 library(MitoHEAR)
@@ -43,7 +53,7 @@ The package has two main functions: **get_raw_counts_allele** and **get_heteropl
 
 ### get_raw_counts_allele
 
-```
+```r
 get_raw_counts_allele(bam_input, path_fasta, cell_names, cores_number = 1) 
 ```
 
@@ -57,10 +67,9 @@ In the same location of the sorted bam file, also the corresponding index bam fi
 Below an example of input using the development version of **MitoHEAR** from GitHub. The example is based on single cell RNA seq mouse embryo data from [Lima *et al.*, Nature Metabolism, 2021 ](https://www.nature.com/articles/s42255-021-00422-7?proof=t):
 
 
-````
 First we download the input bam files (5 samples in the example):
 
-```
+```r
 current_wd <- getwd()
 url <- "https://hmgubox2.helmholtz-muenchen.de/index.php/s/7P9C57RxfKnH5Qx/download/input_bam_files.tar.gz"
 destfile <- paste0(current_wd, "input_bam_files.tar.gz")
@@ -75,13 +84,13 @@ cell_names[1:5]
 ```
 where after_qc is a dataframe with number of rows equal to the number of samples and with columns related to meta data information (i.e. cluster and batch).
 
-```
+```r
 path_to_bam <- paste0(current_wd, "input_bam_files/")
 bam_input <- paste(path_to_bam, cell_names, ".unique.bam", sep = "")
 path_fasta <- system.file("extdata", "Mus_musculus.GRCm38.dna.chromosome.MT.fa", package = "MitoHEAR")
 output_SNP_mt <- get_raw_counts_allele(bam_input, path_fasta, cell_names)
 ```
-````
+
[…]
