### 2022-05-08 ScialdoneLab/MitoHEAR#3 @BatoolMM
Add URL to the description of the package 
I like the Vignettes created with pkgdown (https://github.com/openjournals/joss-reviews/issues/4265), but it's hidden in the README.  Usually, the Vignettes URL is added to the description in the top right of the repo. It makes it more visible and gives a quick summary/overview of the package, especially when you search on GitHub. Below is how I do recommend it:

> Yours show as:

![Screenshot 2022-05-08 at 14 47 26](https://user-images.githubusercontent.com/53487593/167294738-424ee9a2-c64c-47e4-a483-62701b23c0d5.png)

> We recommend including the URL and a few words about the package:

![Screenshot 2022-05-08 at 14 49 43](https://user-images.githubusercontent.com/53487593/167294830-4adffbf9-a75c-46ff-b39b-2c3ce1109d82.png)

Again, this is not necessary but it's a good practice!
