### 2022-05-08 #4265 @BatoolMM
Thank you, @gabrielelubatti, for going through all the comments. I'm satisfied with the changes made to the package and manuscript.

I have added a couple of issues to help with readability and suggestions for fixing the bibliography. My only two cents recommendations:

- Some examples rely on `gam` package to run, but it isn't one of the dependencies. For example, if you want to detect the bases whose heteroplasmy changes along pseudo-time with dpt_test and to plot the trend with plot_dpt (in the README example), you'll need to install `gam` package.

- You might want to rename the titles inside the Vignettes. It becomes more meaningful. For example, rather than `cell_competition_bulk_data_mt_example_notebook`, maybe `Analysing mtDNA cell lines in bulk RNA seq data` or something more meaningful.

- The two main functions have tests, which both run successfully. In terms of the testing, `autotest` is used to perform automatic mutation testing of the package, where you can see if the test implemented is working as anticipated.  In other words, it's used to verify the test you implemented in `testthat`, if that makes sense. `autotest_package ("MitoHEAR")` returns NULL, which indicates that all tests passed successfully.

This concludes my review. I'm happy with the current version of the package and manuscript - I had fun reading and reviewing it.
