+++ vignettes/lineage_tracing_example_notebook.Rmd
@@ -1,8 +1,8 @@
 ---
-title: "lineage_tracing_example_notebook"
+title: "Lineage tracing with mtDNA heteroplasmy in single cell RNA seq mouse data"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{lineage_tracing_example_notebook}
+  %\VignetteIndexEntry{Lineage tracing with mtDNA heteroplasmy in single cell RNA seq mouse data}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
