+++ Draft/paper.bib
@@ -1,5 +1,5 @@
 @article {Miller2021,
-	author = {Miller, Tyler E. and Lareau, Caleb A. and Verga, Julia A. and Ssozi, Daniel and Ludwig, Leif S. and Farran, Chadi El and Griffin, Gabriel K. and Lane, Andrew A. and Bernstein, Bradley E. and Sankaran, Vijay G. and van Galen, Peter},
+	author = {Miller, Tyler E. and Lareau, Caleb A. and Verga, Julia A. and Ssozi, Daniel and Ludwig, Leif S. and Farran, Chadi El and Griffin, Gabriel K. and Lane, Andrew A. and Bernstein, Bradley E. and Sankaran, Vijay G. and {van Galen}, Peter},
 	title = {Mitochondrial variant enrichment from high-throughput single-cell RNA-seq resolves clonal populations},
 	elocation-id = {2021.03.08.434450},
 	year = {2021},
@@ -59,13 +59,13 @@ keywords = {lineage tracing, single cell genomics, somatic mutations, mitochondr
 	doi = {10.1038/s42255-021-00422-7},
 	URL = {https://www.nature.com/articles/s42255-021-00422-7},
 	eprint = {https://www.nature.com/articles/s42255-021-00422-7.pdf},
-	journal = {Nat Metab}
+	journal = {Nature Metabolism}
 }
 
 @article{Nissanka2020,
 author = {Nissanka, Nadee and Moraes, Carlos T},
 title = {Mitochondrial DNA heteroplasmy in disease and targeted nuclease-based therapeutic approaches},
-journal = {EMBO Rep},
+journal = {EMBO Reports},
 volume = {21},
 number = {3},
 pages = {e49612},
@@ -80,16 +80,16 @@ year = {2020}
 	author = {Stewart, J., Chinnery, P.},
 	title = {The dynamics of mitochondrial DNA heteroplasmy: implications for human health and disease},
 	year = {2015},
-	doi = {doi.org/10.1038/nrg3966},
-	journal = {Nat Rev Genet}
+	doi = {10.1038/nrg3966},
+	journal = {Nature Reviews Genetics}
 }
 
 @article {Floros2019,
-	author = {Floros VI, Pyle A, Dietmann S, Wei W, Tang WCW, Irie N, Payne B, Capalbo A, Noli L, Coxhead J, Hudson G, Crosier M, Strahl H, Khalaf Y, Saitou M, Ilic D, Surani MA, Chinnery PF.},
+	author = {Floros, VI and Pyle, A and Dietmann, S and Wei, W and Tang, WCW and Irie, N and Payne, B and Capalbo, A and Noli, L and Coxhead, J and Hudson, G and Crosier, M and Strahl, H and Khalaf, Y and Saitou, M and Ilic, D and Surani, MA and Chinnery, PF},
 	title = {Segregation of mitochondrial DNA heteroplasmy through a developmental genetic bottleneck in human embryos},
 	year = {2019},
 	doi = {10.1038/s41556-017-0017-8},
-	journal = {Nat Cell Biol}
+	journal = {Nature Cell Biology}
 }
 @article{Calabrese2014,
     author = {Calabrese, Claudia and Simone, Domenico and Diroma, Maria Angela and Santorsola, Mariangela and Guttà, Cristiano and Gasparre, Giuseppe and Picardi, Ernesto and Pesole, Graziano and Attimonelli, Marcella},
