### 2022-05-30 #4265 @csoneson
@gabrielelubatti I have sent a PR with some minor fixes to the manuscript/bibliography. Could you please check and merge if you agree, and then generate a new proof with `@editorialbot generate pdf`? I also noticed that some of the preprints cited in the manuscript are now peer-reviewed and published - could you also update those citations if applicable?

Then, the next steps are as follows:
- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
