+++ Draft/paper.bib
@@ -1,13 +1,12 @@
-@article {Miller2021,
+@article {Miller2022,
 	author = {Miller, Tyler E. and Lareau, Caleb A. and Verga, Julia A. and Ssozi, Daniel and Ludwig, Leif S. and Farran, Chadi El and Griffin, Gabriel K. and Lane, Andrew A. and Bernstein, Bradley E. and Sankaran, Vijay G. and {van Galen}, Peter},
 	title = {Mitochondrial variant enrichment from high-throughput single-cell RNA-seq resolves clonal populations},
 	elocation-id = {2021.03.08.434450},
-	year = {2021},
-	doi = {10.1101/2021.03.08.434450},
-	publisher = {Cold Spring Harbor Laboratory},
-	URL = {https://www.biorxiv.org/content/early/2021/03/09/2021.03.08.434450},
-	eprint = {https://www.biorxiv.org/content/early/2021/03/09/2021.03.08.434450.full.pdf},
-	journal = {bioRxiv}
+	year = {2022},
+	doi = {10.1038/s41587-022-01210-8},
+	URL = {https://www.nature.com/articles/s41587-022-01210-8},
+	eprint = {https://www.nature.com/articles/s41587-022-01210-8.pdf},
+	journal = {Nature Biotechnology}
 }
 
 
