+++ Draft/paper.bib
@@ -1,7 +1,6 @@
 @article {Miller2022,
 	author = {Miller, Tyler E. and Lareau, Caleb A. and Verga, Julia A. and Ssozi, Daniel and Ludwig, Leif S. and Farran, Chadi El and Griffin, Gabriel K. and Lane, Andrew A. and Bernstein, Bradley E. and Sankaran, Vijay G. and {van Galen}, Peter},
 	title = {Mitochondrial variant enrichment from high-throughput single-cell RNA-seq resolves clonal populations},
-	elocation-id = {2021.03.08.434450},
 	year = {2022},
 	doi = {10.1038/s41587-022-01210-8},
 	URL = {https://www.nature.com/articles/s41587-022-01210-8},
@@ -39,16 +38,14 @@ keywords = {lineage tracing, single cell genomics, somatic mutations, mitochondr
 }
 
 
-@article {Prashant2020,
+@article {Prashant2021,
 	author = {Prashant, NM and Alomran, Nawaf and Chen, Yu and Liu, Hongyu and Bousounis, Pavlos and Movassagh, Mercedeh and Edwards, Nathan and Horvath, Anelia},
-	title = {SCReadCounts: Estimation of cell-level SNVs from scRNA-seq data},
-	elocation-id = {2020.11.23.394569},
+	title = {SCReadCounts: estimation of cell-level SNVs expression from scRNA-seq data},
 	year = {2021},
-	doi = {10.1101/2020.11.23.394569},
-	publisher = {Cold Spring Harbor Laboratory},
-	URL = {https://www.biorxiv.org/content/early/2021/06/12/2020.11.23.394569},
-	eprint = {https://www.biorxiv.org/content/early/2021/06/12/2020.11.23.394569.full.pdf},
-	journal = {bioRxiv}
+	doi = {10.1186/s12864-021-07974-8},
+	URL = {https://bmcgenomics.biomedcentral.com/articles/10.1186/s12864-021-07974-8},
+	eprint = {https://bmcgenomics.biomedcentral.com/track/pdf/10.1186/s12864-021-07974-8.pdf},
+	journal = {BMC Genomics}
 }
 
 @article {Lima2020,
