### 2022-05-31 #4265 @gabrielelubatti
Dear @csoneson  thanks a lot for the information!
I updated some citations in the paper and merged your PR. I also archived the software on Zenodo (version v1.0.1) with DOI 10.5281/zenodo.6598821.
Let me know If something is unclear or if I need to do something else.
Best
