### 2022-05-31 #4265 @csoneson
Thanks @gabrielelubatti - I noticed one more thing that I missed before, sorry: the reference for Stewart (2015) seems misformatted, since it only shows one author in the reference list instead of the two that should be there. Could you fix that?
