+++ Draft/paper.bib
@@ -73,7 +73,7 @@ year = {2020}
 }
 
 @article {Stewart2015,
-	author = {Stewart, J., Chinnery, P.},
+	author = {Stewart, J and Chinnery, P},
 	title = {The dynamics of mitochondrial DNA heteroplasmy: implications for human health and disease},
 	year = {2015},
 	doi = {10.1038/nrg3966},
