+++ Draft/paper.md
@@ -51,11 +51,11 @@ Among the downstream analyses implemented in the package are:
 * Plotting functions for the visualisation of heteroplasmy and the corresponding allele frequency values among cells.
 * Unsupervised hierarchical clustering of cells based on a distance matrix defined from the angular distance of allele frequencies that could be relevant for lineage tracing analysis [@LUDWIG20191325] (**Figure 3**).
 
-![Example of an output plot generated by MitoHEAR showing heteroplasmy values at a given position estimated from single cells in three clusters indicated on the x-axis. Data from @Lima2020. \label{fig:flowshart}](docs/img/paper_fig_1.png){ width=60%}
+![Example of an output plot generated by MitoHEAR showing heteroplasmy values at a given position estimated from single cells in three clusters indicated on the x-axis. Data from @Lima2020. \label{fig:flowshart1}](docs/img/paper_fig_1.png){ width=60%}
 
-![Example of an output figure generated by MitoHEAR where the heteroplasmy is plotted as a function of the pseudo-time coordinate of each cell. Cells are classified into three clusters. The heteroplasmy shows a statistically significant change along the pseudo-time, as indicated by the adjusted p-value reported at the top, which is computed by a generalised additive model fit. Data from @Lima2020. \label{fig:flowshart}](docs/img/paper_fig_2.png){ width=60%}
+![Example of an output figure generated by MitoHEAR where the heteroplasmy is plotted as a function of the pseudo-time coordinate of each cell. Cells are classified into three clusters. The heteroplasmy shows a statistically significant change along the pseudo-time, as indicated by the adjusted p-value reported at the top, which is computed by a generalised additive model fit. Data from @Lima2020. \label{fig:flowshart2}](docs/img/paper_fig_2.png){ width=60%}
 
-![Unsupervised hierarchical clustering of cells based on a distance matrix defined from the angular distance of allele frequencies. The data shown is bulk RNA-seq mouse data from two mtDNA cell lines labelled *Loser* and *Winner*. Data from @Lima2020. \label{fig:flowshart}](docs/img/paper_fig_3.png){ width=60%}
+![Unsupervised hierarchical clustering of cells based on a distance matrix defined from the angular distance of allele frequencies. The data shown is bulk RNA-seq mouse data from two mtDNA cell lines labelled *Loser* and *Winner*. Data from @Lima2020. \label{fig:flowshart3}](docs/img/paper_fig_3.png){ width=60%}
 
 The package has been used in a recently published paper [@Lima2020], where we revealed that cells with higher levels of heteroplasmy are eliminated by cell competition in mouse embryos and are characterised by specific gene expression patterns.
 
