### 2022-06-01 #4265 @gabrielelubatti
Hello @csoneson I should have renamed the labels of the figures.
