### 2022-06-01 #4265 @danielskatz
👋 @gabrielelubatti - I'm the AEiC on duty this week, handing the rest of the publication process for your paper. I have some suggested changes in https://github.com/ScialdoneLab/MitoHEAR/pull/6 - please merge this or let me know what you disagree with, then we can proceed.
